LetsCookOurResume/requirements.txt
alembic==1.13.2
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
babel==2.16.0
bcrypt==4.2.0
billiard==3.6.4.0
blinker==1.8.2
boto3==1.35.78
botocore==1.35.78
celery==5.2.7
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
cssmin==0.2.0
Deprecated==1.2.14
distlib==0.3.8
exceptiongroup==1.2.2
filelock==3.16.0
flake8==6.0.0
Flask==2.2.3
Flask-Admin==1.6.1
Flask-Assets==2.0
flask-babel==3.1.0
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-DebugToolbar==0.13.1
Flask-Limiter==3.3.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-QueryInspect==0.1.2
Flask-RESTful==0.3.9
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
flower==1.2.0
greenlet==3.1.0
gunicorn==23.0.0
humanize==4.10.0
idna==3.8
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsmin==3.0.1
kombu==5.4.1
limits==3.13.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mock==5.0.2
ordered-set==4.1.0
packaging==24.1
passlib==1.7.4
platformdirs==3.11.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pycodestyle==2.10.0
pydantic==2.9.1
pydantic_core==2.23.3
pyflakes==3.0.1
Pygments==2.18.0
PyJWT==2.10.1
pypdf==4.3.1
PyPDF2==3.0.1
pytest==7.3.1
pytest-cov==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.8.1
s3transfer==0.10.4
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.34
tomli==2.0.1
tornado==6.3.1
typing_extensions==4.12.2
urllib3==2.2.3
vine==5.1.0
virtualenv==20.22.0
wcwidth==0.2.13
webassets==2.0
Werkzeug==2.2.2
wrapt==1.16.0
WTForms==3.1.2
